pyproject.toml
[tool.poetry]
name = "backend_django"
version = "0.1.0"
description = ""
authors = ["kpaveliev <kpaveliev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
djoser = "^2.1.0"
Pillow = "^9.2.0"
drf-spectacular = "^0.23.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
